+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Added:
 * If 'Show Race Name In Server Lobby' is enabled, Quick Races and Custom Races with no specified name now show the track name in the Server Lobby.
 * A global option to switch speed units from Km/h to MPH for people who want to use the correct measurement system.
 * "Force Virtual Mirror" now defaults to on in all race setup forms.
+* Admin Control Panel on the Live Timing page. Allows admins to send server wide messages, advance to the next/restart session, kick users and other admin commands!
 * You can now re-order Championship Events! Drag them around on the page using the top bar of the Championship Event.
 
 Fixes: